kernelci.db: rename from kernelci.data
Rename kernelci.data into kernelci.db to avoid ambibuity between data
files and database abstraction layers.  Update places where
kernelci.data was being used accordingly.
