+++ kci_data
@@ -23,8 +23,8 @@ import sys
 
 from kernelci.cli import Args, Command, parse_opts
 import kernelci.build
-import kernelci.config.data
-import kernelci.data
+import kernelci.config.db
+import kernelci.db
 
 
 class cmd_validate(Command):
@@ -67,7 +67,7 @@ class cmd_submit(Command):
         else:
             with open(args.data_file, 'r') as json_file:
                 data = json.load(json_file)
-        db = kernelci.data.get_db(config, args.db_token)
+        db = kernelci.db.get_db(config, args.db_token)
         return db.submit(data, args.verbose)
 
 
@@ -80,7 +80,7 @@ class cmd_submit_build(Command):
         config = config_data['db_configs'][args.db_config]
         install = kernelci.build.Step.get_install_path(args.kdir, args.output)
         meta = kernelci.build.Metadata(install)
-        db = kernelci.data.get_db(config, args.db_token)
+        db = kernelci.db.get_db(config, args.db_token)
         return db.submit_build(meta, args.verbose)
 
 
@@ -93,7 +93,7 @@ class cmd_submit_test(Command):
         config = config_data['db_configs'][args.db_config]
         with open(args.data_file, 'r') as json_file:
             data = json.load(json_file)
-        db = kernelci.data.get_db(config, args.db_token)
+        db = kernelci.db.get_db(config, args.db_token)
         return db.submit_test(data, args.verbose)
